# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for TechTask Application",
        "title": "TechTask App api",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/currency/save/{date}": {
            "get": {
                "description": "создание r_currency на основе данных с api",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "create a new rcurrencies"
                ],
                "summary": "RCurrency create",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SuccessResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/currency/{date}/{code}": {
            "get": {
                "description": "Получение списка r_currency, pathvariable code  не обязательный",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List of rcurrencies"
                ],
                "summary": "RCurrencies list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.RCurrency"
                            }
                        }
                    },
                    "204": {
                        "description": "No content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "model.RCurrency": {
            "type": "object",
            "properties": {
                "a_date": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "model.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.RCurrency"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}